% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf-package.R
\docType{package}
\name{dcf-package}
\alias{dcf}
\alias{dcf-package}
\title{dcf: dcf: Data Collection Framework}
\description{
Establishes and operates within a framework to collect data in a repeatable way.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://DISSC-yale.github.io/dcf}
  \item \url{https://github.com/DISSC-yale/dcf}
  \item Report bugs at \url{https://github.com/DISSC-yale/dcf/issues}
}

}
\author{
\strong{Maintainer}: Micah Iserman \email{micah.iserman@gmail.com}

Other contributors:
\itemize{
  \item Yale Data-Intensive Social Science Center [funder, copyright holder]
}

}
\keyword{internal}
