% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_add_bundle.R
\name{dcf_add_bundle}
\alias{dcf_add_bundle}
\title{Adds a Bundle Project}
\usage{
dcf_add_bundle(
  name,
  project_dir = ".",
  source_files = NULL,
  open_after = interactive(),
  use_git = TRUE,
  use_workflow = FALSE
)
}
\arguments{
\item{name}{Name of the bundle}

\item{project_dir}{Path to the Data Collection Framework project.}

\item{source_files}{A list or character vector, with names as paths to standard files
form source projects (relative to the project's data directory), and distribution file
names as entries. This associates input with output files, allowing for calculation of
a source state, and metadata inheritance from source files.}

\item{open_after}{Logical; if \code{FALSE}, will not open the project.}

\item{use_git}{Logical; if \code{TRUE}, will initialize a git repository.}

\item{use_workflow}{Logical; if \code{TRUE}, will add a GitHub Actions workflow.}
}
\value{
Nothing; creates default files and directories.
}
\description{
Establishes a new data bundle project, used to prepare outputs from standardized datasets.
}
\section{Project Definition}{


The \strong{\code{process.json}} file defines the project with some initial attributes:
\itemize{
  \item \code{type} Always \code{bundle} to define this as a bundle project.
  \item \code{name} Name of the project.
  \item \code{scripts} List of script definitions.
  \item \code{source_files} A character array of paths to other files
    used within the scripts, relative to the overall project's \code{data} directory.
  \item \code{standard_state} State of the \code{source_files}: A list
    with keys as the file paths, relative to the overall project root, and values
    as the MD5 hash of those files.
  \item \code{dist_state} State of the \code{dist} directory: A list
    with keys as the file paths, relative to the overall project root, and values
    as the MD5 hash of those files.
  \item \code{checked} Timestamp when the project was last checked with \code{\link{dcf_check}}.
  \item \code{check_results} Results of the last check.
}

Each \strong{\code{scripts}} entry points to a script to be run, with one default:
\itemize{
  \item \code{path} path to the script, relative to this project's root.
  \item \code{last_run} Timestamp of the last processing.
  \item \code{run_time} How long the script took to run last, in milliseconds.
  \item \code{last_status} Status of the last run; a list with entries for
    \code{success} (logical) and \code{log} (output of the script).
}
}

\section{Project Files}{


Within a bundle project, there are two files to edits:
\itemize{
  \item \strong{\code{build.R}}: This is the primary script, which is automatically rerun.
    It should read data from the \code{standard} directory of source projects,
    and write to it's own \code{dist} directory.
  \item \strong{\code{measure_info.json}}: This should list all non-ID variable names
    in the data files within \code{dist}. These will inherit the standard measure info
    if found in the source projects referred to in \code{source_files}.
    If the \code{dist} name is different, but should still inherit standard measure info,
    a \code{source_id} entry with the original measure ID will be used to identify the original
    measure info.
    See \code{\link{dcf_measure_info}}.
}
}

\examples{
project_dir <- file.path(tempdir(), "temp_project")
dcf_init("temp_project", dirname(project_dir))
dcf_add_bundle("bundle_name", project_dir)
list.files(file.path(project_dir, "data/bundle_name"))

}
