% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_check.R
\name{dcf_check}
\alias{dcf_check}
\title{Check Data Projects}
\usage{
dcf_check(names = NULL, project_dir = ".", verbose = TRUE)
}
\arguments{
\item{names}{Name or names of projects.}

\item{project_dir}{Path to the Data Collection Framework project.}

\item{verbose}{Logical; if \code{FALSE}, will not print status messages.}
}
\value{
A list with an entry for each source, containing a character vector
  including any issue codes:
\itemize{
  \item \code{not_compressed}: The file does not appear to be compressed.
  \item \code{cant_read}: Failed to read the file in.
  \item \code{geography_nas}: The file's \code{geography} column contains NAs.
  \item \code{geography_dropped}: The file's \code{geography} column has levels
    dropped from previous versions.
  \item \code{time_nas}: The file's \code{time} column contains NAs.
  \item \code{missing_info: {column_name}}: The file's indicated column does not have
    a matching entry in \code{measure_info.json}.
  \item \code{dropped_measure: {column_name}}: The file's indicated column is not present
    when it was previously.
  \item \code{type_changed: {column_name}}: The file's indicated column's type changed
    from the previous version.
  \item \code{levels_changed: {column_name}}: The file's indicated column's levels have
    all changed from the previous version.
}
}
\description{
Check the data files and measure info of subprojects.
}
\examples{
\dontrun{

  # run from within a project or sub-project
  dcf_check()

  # run from within a project on a single sub-project
  dcf_check("gtrends")

  # run outside of a project on a single sub-project
  dcf_check("gtrends", "../path/to/project")
}
}
