% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_download_wisqars.R
\name{dcf_download_wisqars}
\alias{dcf_download_wisqars}
\title{Download CDC WISQARS Reports}
\usage{
dcf_download_wisqars(
  file,
  fatal_outcome = TRUE,
  brain_injury_only = FALSE,
  year_start = 2018,
  year_end = year_start,
  geography = "00",
  intent = "all",
  disposition = "all",
  mechanism = if (fatal_outcome) 20810 else 3000,
  group_ages = NULL,
  age_min = 0,
  age_max = 199,
  sex = "all",
  race = "all",
  race_reporting = "single",
  ethnicity = "all",
  YPLL = 65,
  metro = NULL,
  group_by = NULL,
  include_total = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{file}{File to save the report to (\code{csv} or \code{parquet}).}

\item{fatal_outcome}{Logical; if \code{FALSE}, will return non-fatal results.}

\item{brain_injury_only}{Logical; if \code{TRUE}, will return only traumatic brain injury results.}

\item{year_start}{Earliest year to include.}

\item{year_end}{Latest year to include.}

\item{geography}{State or region code.}

\item{intent}{Intent ID or name:
\tabular{lll}{
  \code{0} \tab \code{all} \tab All \cr
  \code{1} \tab \code{unintentional} \tab Unintentional \cr
  \code{2} \tab \code{suicide} \tab Suicide \cr
  \code{3} \tab \code{homicide} \tab Homicide \cr
  \code{4} \tab \code{homicide_legal} \tab Homicide and Legal Intervention \cr
  \code{5} \tab \code{undetermined} \tab Undetermined \cr
  \code{6} \tab \code{legal} \tab Legal Intervention \cr
  \code{8} \tab \code{violence} \tab Violence-related \cr
}}

\item{disposition}{Patient disposition given nonfatal: one or multiple of \code{all} (0),
\code{treated} (1; treated and released), \code{transfered} (2), \code{hospitalized} (3), or
\code{observed} (4; observed/left AMA/unknown).}

\item{mechanism}{Mechanism code; default to \code{20810} (all injury).
Other codes appear in the URL in the \code{m} parameter when submitting the filter.}

\item{group_ages}{Logical; if \code{FALSE}, will not group ages into 5-year bins.}

\item{age_min}{Youngest age to include.}

\item{age_max}{Oldest age to include.}

\item{sex}{Sex groups to include: one or multiple of \code{all} (0), \code{male} (1),
\code{female} (2), or \code{unknown} (3)..}

\item{race}{Race groups to include: one or multiple of \code{all} (0), \code{white} (1),
\code{black} (2), \code{aa} (3; American Indian or Alaska Native), \code{asian} (4),
\code{pi} (5; Hawaiian Native or Pacific Islander), \code{more} (6; more than one race).
These levels apply if \code{race_reporting} is \code{single} (default) -- provide
these by index for other \code{race_reporting} levels.}

\item{race_reporting}{How to group race groups, between \code{none} (0), \code{bridge} (1),
\code{single} (2), or \code{aapi} (3).}

\item{ethnicity}{Which ethnic groups to include: one or multiple of \code{all} (0),
\code{non_hispanic} (1), \code{hispanic} (2), or \code{unknown} (3).}

\item{YPLL}{Age to use when calculating Years of Potential Life Lost.}

\item{metro}{Region type filter: \code{TRUE} for only metropolitan / urban, or \code{FALSE}
for only non-metropolitan / rural. Will include all region types if \code{NULL} (default).}

\item{group_by}{One or more variables to group by. These are uppercased and sometimes
abbreviated or encoded; see the \code{r1} through \code{r4} URL parameters.}

\item{include_total}{Logical; if \code{FALSE}, will not include totals.}

\item{verbose}{Logical; if \code{FALSE}, will not display status messages.}
}
\value{
A list containing the parameters of the request. The returned data are written to \code{file}.
}
\description{
Download reports data from the CDC's
\href{https://wisqars.cdc.gov/reports}{Web-based Injury Statistics Query and Reporting System}.
}
\examples{
file <- "../../../wisqars.csv.xz"
if (file.exists(file)) {
  dcf_download_wisqars(file, verbose = FALSE)
  vroom::vroom(file)
}
}
