% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_overview_app.R
\name{dcf_efish_app}
\alias{dcf_efish_app}
\title{dcf_efish_app: A Shiny app for displaying e-fishing data}
\usage{
dcf_efish_app()
}
\description{
This app displays e-fishing data from known DCF efishing sites.
The primary use is to examine if all expected sites have been fished and
identify missing sites and also duplicates
}
\examples{
\dontrun{
library(DCF)
dcf_efish_app()
}
}
