% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dcf_get_efish_data}
\alias{dcf_get_efish_data}
\title{Get electrofishing data for a river for one species one year}
\usage{
dcf_get_efish_data(
  river,
  year = as.numeric(format(Sys.Date(), "\%Y")),
  year2 = NULL,
  wanted_sites = dcf_known_efish_sites(river)
)
}
\arguments{
\item{river}{character string with name of river.}

\item{year}{numeric year to query or the string "all". Default current year}

\item{year2}{optional numeric to return results between year and year2}

\item{species}{character string indication what species to query. Default "Lax".}

\item{sites}{a data frame with xkoorlok and ykoorlok}
}
\value{
Return a data frame with electro fishing results. If "fiskedat" is NA the site was not
fished in the time period.
}
\description{
Get electrofishing data for a river for one species one year
}
\examples{
\dontrun{
res <- dcf_get_efish_data("Vindelälven", species = "Lax", year= 2018)
head(res)
}
}
