% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_init.R
\name{dcf_init}
\alias{dcf_init}
\title{Initialize a Data Collection Project}
\usage{
dcf_init(
  name,
  base_dir = ".",
  data_dir = "data",
  github_account = "",
  branch = "main",
  repo_name = name,
  use_git = TRUE,
  open_after = FALSE
)
}
\arguments{
\item{name}{Name of the source. Defaults to the current directory name.}

\item{base_dir}{Path to the parent of the project directory (where the \code{name}
directory should be created). If \code{name} is not specified, will treat the current
directory as \code{name}, and \code{".."} as \code{base_dir}.}

\item{data_dir}{Name of the directory to store projects in, relative to \code{base_dir}.}

\item{github_account}{Name of the GitHub account that will host the repository.}

\item{branch}{Name of the repository's branch.}

\item{repo_name}{Name of the repository.}

\item{use_git}{Logical; if \code{TRUE}, will initialize a git repository.}

\item{open_after}{Logical; if \code{TRUE}, will open the project in a new RStudio instance.}
}
\value{
Nothing; creates default files and directories.
}
\description{
Establishes a new data collection framework project.
}
\section{Data Collection Project}{


A data collection project starts with a \code{settings.json} file, which
specifies where source and bundle projects live (a \code{data} subdirectory by default).

The bulk of the project will then be in the source and bundle projects, as created
by the \code{\link{dcf_add_source}} and \code{\link{dcf_add_bundle}}.

Once these sub-projects are in place, they can be operated over by the
\code{\link{dcf_build}}, which processes each sub-project using
\code{\link{dcf_process}}, and checks them with \code{\link{dcf_check}},
resulting in a report.
}

\examples{
base_dir <- tempdir()
dcf_init("project_name", base_dir)
list.files(file.path(base_dir, "project_name"))

}
