% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_process.R
\name{dcf_process}
\alias{dcf_process}
\title{Run Data Project Processes}
\usage{
dcf_process(
  name = NULL,
  project_dir = ".",
  ingest = TRUE,
  is_auto = FALSE,
  force = FALSE,
  clear_state = FALSE
)
}
\arguments{
\item{name}{Name of a source project to process. Will}

\item{project_dir}{Path to the project directory. If not specified, and being called
from a source project, this will be assumed to be two steps back from the working directory.}

\item{ingest}{Logical; if \code{FALSE}, will re-process standardized data without running
ingestion scripts. Only applies to source projects.}

\item{is_auto}{Logical; if \code{TRUE}, will skip process scripts marked as manual.}

\item{force}{Logical; if \code{TRUE}, will ignore process frequencies
(will run scripts even if recently run).}

\item{clear_state}{Logical; if \code{TRUE}, will clear stored states before processing.}
}
\value{
A list with processing results:
\itemize{
  \item \code{timings}: How many seconds the scripts took to run.
  \item \code{logs}: The captured output of the scripts.
}
Each entry has an entry for each project.

A `datapackage.json` file is also created / update in each source's `standard` directory.
}
\description{
Operates over data source and bundle projects, optionally running the source
ingest scripts, then collecting metadata.
}
\examples{
\dontrun{
  # run from a directory containing a `data` directory containing the source
  dcf_process("source_name")

  # run without executing the ingestion script
  dcf_process("source_name", ingest = FALSE)
}
}
