% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_process_epic_staging.R
\name{dcf_process_epic_staging}
\alias{dcf_process_epic_staging}
\title{Process Epic Staging Files}
\usage{
dcf_process_epic_staging(
  staging_dir = "raw/staging",
  out_dir = "raw",
  verbose = TRUE,
  cleanup = TRUE,
  ...
)
}
\arguments{
\item{staging_dir}{Directory containing the staging files.}

\item{out_dir}{Directory to write new raw files to.}

\item{verbose}{Logical; if \code{FALSE}, will not show status messages.}

\item{cleanup}{Logical; if \code{FALSE}, will not remove staging files after being processed.}

\item{...}{Passes additional arguments to \code{\link{dcf_read_epic}}.}
}
\value{
\code{NULL} if no staging files are found.
  Otherwise, a list with entries for \code{data} and \code{metadata}.
  Each of these are lists with entries for each recognized standard name,
  with potentially combined outputs similar to \code{\link{dcf_read_epic}}
}
\description{
Process Epic staging files, lightly standardizing them and moving them to raw.
}
\examples{
\dontrun{
  # run from a source project
  dcf_process_epic_staging()
}

}
