% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_process_record.R
\name{dcf_process_record}
\alias{dcf_process_record}
\title{Interact with a Process File}
\usage{
dcf_process_record(path = "process.json", updated = NULL)
}
\arguments{
\item{path}{Path to the process JSON file.}

\item{updated}{An update version of the process definition. If specified, will
write this as the new process file, rather than reading any existing file.}
}
\value{
The process definition of the source project.
}
\description{
Read or update the current process file.
}
\details{
See the \href{https://dissc-yale.github.io/dcf/articles/standards.html#scripts}{script standards}
for examples of using this within a sub-project script.
}
\examples{
epic_process_file <- "../../../pophive/pophive_demo/data/epic/process.json"
if (file.exists(epic_process_file)) {
  dcf_process_record(epic_process_file)
}
}
