% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_read_epic.R
\name{dcf_read_epic}
\alias{dcf_read_epic}
\title{Read Epic Cosmos Data}
\usage{
dcf_read_epic(path, path_root = ".", standard_names = NULL)
}
\arguments{
\item{path}{Path to the file.}

\item{path_root}{Directory containing \code{path}, if it is not full.}

\item{standard_names}{A character vector with standard dataset names as names, and
fixed patterns to search for in the metadata as values (in lowercase; e.g.,
\code{c(condition = "condition name")}).
These take precedence over the existing set of standard names, so make sure
the pattern is sufficiently specific to the target dataset.}
}
\value{
A list with \code{data.frame} entries for \code{metadata} and \code{data}.
}
\description{
Read in metadata and data from an Epic Cosmos file.
}
\examples{
# write an example file
path <- tempfile(fileext = ".csv")
raw_lines <- c(
  "metadata field,metadata value,",
  ",,",
  ",Measures,Value Name",
  "Year,Measure 1,",
  "2020,m1,1",
  ",m2,2",
  "2021,m1,3",
  ",m2,4"
)
writeLines(raw_lines, path)

# read it in
dcf_read_epic(basename(path), dirname(path))

}
