% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_report.R
\name{dcf_report}
\alias{dcf_report}
\title{Retrieve Project Report}
\usage{
dcf_report(
  project = "dissc-yale/pophive_demo",
  branch = "main",
  commit = NULL,
  provider = "https://github.com",
  cache = tempdir(),
  refresh = FALSE
)
}
\arguments{
\item{project}{Path to a local project, or the GitHub account and repository name
(\code{"{account_name}/{repo_name}"}) of a remote project.}

\item{branch}{Name of the remote repository branch.}

\item{commit}{Commit hash; overrides \code{branch}.}

\item{provider}{Base URL of the remote repository provider.}

\item{cache}{Directory to store retrieved report in (at \code{{cache}/{project}/report.json.gz}).}

\item{refresh}{Logical; if \code{TRUE}, will always retrieve a fresh copy of the report,
even if a copy exists in \code{cache}.}
}
\value{
A data collection project report:
\tabular{ll}{
  \code{date} \tab Timestamp when the report was created. \cr
  \code{settings} \tab The project's settings file. \cr
  \code{source_times} \tab
    A list with entries for each subproject, containing the number of milliseconds
    it took to run the project's scripts. \cr
  \code{issues} \tab
    A list with entries for each subproject, containing issues flagged by
    \code{\link{dcf_check}}, within a list with \code{data} and/or \code{measure} entries,
    containing character vectors of issue labels. \cr
  \code{logs} \tab
    A list with entries for each subproject, containing the logged output of their scripts. \cr
  \code{metadata} \tab
    A list with entries for each subproject, containing the datapackage of their output,
    as created by \code{\link{dcf_measure_info}}. \cr
  \code{processes} \tab
    A list with entries for each subproject, containing their process definitions
    (see \code{\link{dcf_add_source}} and/or \code{\link{dcf_add_bundle}}). \cr
}
}
\description{
Retrieve the report file from a local or remote project.
}
\examples{
report <- dcf_report("dissc-yale/pophive_demo")
report$date
jsonlite::toJSON(report$settings, auto_unbox = TRUE, pretty = TRUE)
}
