% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_to_health_region.R
\name{dcf_to_health_region}
\alias{dcf_to_health_region}
\title{Map States to Health Regions}
\usage{
dcf_to_health_region(geoids, prefix = "Region ")
}
\arguments{
\item{geoids}{Character vector of GEOIDs.}

\item{prefix}{A prefix to add to region IDs.}
}
\value{
A vector of Health Region names the same length as \code{geoids}.
}
\description{
Maps state FIPS state numeric codes to Human Health Service regions.
}
\examples{
dcf_to_health_region(c("01", "01001", "02", "02001"))
}
