% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_update_lock.R
\name{dcf_update_lock}
\alias{dcf_update_lock}
\title{Update renv.lock}
\usage{
dcf_update_lock(project_dir = ".", refresh = TRUE)
}
\arguments{
\item{project_dir}{Directory of the Data Collection project.}

\item{refresh}{Logical; if \code{FALSE}, will update an existing
\code{renv.lock} file, rather than recreating it.}
}
\value{
Nothing; writes an \code{renv.lock} file.
}
\description{
Updates the \code{renv.lock} file with dependencies found in project scripts.
}
\examples{
\dontrun{
  dcf_update_lock()
}
}
