% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_variables.R
\name{dcf_variables}
\alias{dcf_variables}
\title{View Project Variables}
\usage{
dcf_variables(project = ".", exclude = c("geography", "time", "age"), ...)
}
\arguments{
\item{project}{Path to a local project, or the GitHub account and repository name
(\code{"{account_name}/{repo_name}"}) of a remote project.
Or a report as returned from \code{\link{dcf_report}}.}

\item{exclude}{A character vector of variable names to exclude from the list (usually ID columns).}

\item{...}{Additional arguments passed to \code{\link{dcf_report}}.}
}
\value{
A tibble containing variables:
\tabular{ll}{
  \code{name} \tab Name of the variable, as it appears in the data file. \cr
  \code{type} \tab The value's storage type. \cr
  \code{n} \tab Number of non-missing observations within the file. \cr
  \code{duplicates} \tab Number of duplicated values within the file. \cr
  \code{missing} \tab Number of missing values within the file. \cr
  \code{project_type} \tab The project type, between \code{source} and \code{bundle}. \cr
  \code{data_format} \tab The orientation of the data, between \code{wide} and \code{tall}. \cr
  \code{file} \tab The file containing the variable; a path relative to the project root. \cr
  \code{short_name} \tab Short name, if included in measure info. \cr
  \code{long_name} \tab Long name, if included in measure info. \cr
  \code{short_decription} \tab Short description, if included in measure info. \cr
  \code{long_description} \tab Long description, if included in measure info. \cr
  \code{measure_type} \tab Higher-level description of type than storage type
    (e.g., \code{count} versus \code{integer}), if included in measure info. \cr
  \code{unit} \tab How a single value should be interpreted
    (e.g., \code{per 100k people} for a rate per 100k people), if included in measure info. \cr
  \code{time_resolution} \tab The measure's collection frequency, if included in measure info. \cr
  \code{category} \tab The measure's category, if included in measure info. \cr
  \code{subcategory} \tab The measure's subcategory, if included in measure info. \cr
}
}
\description{
Get a list of measures (a codebook) that are included in a data collection project.
}
\examples{
dcf_variables("dissc-yale/pophive_demo")
}
\seealso{
Other data user interface functions: 
\code{\link{dcf_data}()},
\code{\link{dcf_report}()}
}
\concept{data user interface functions}
