% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efish_sites}
\alias{efish_sites}
\alias{WGBAST_rivers}
\alias{WGNAS_rivers}
\title{Electofishing sites monitored in the DCF-program (WGBAST and WGNAS)}
\format{
WGBAST_rivers: A vector of  names of Swedish WGBAST rivers. The names can be used to select data from
`efish_sites`.

WGNAS_rivers: A vector of  names of Swedish WGNAS rivers. The names can be used to select data from
`efish_sites`.

efish_sites: A list of data frames, one for each river. Each item have is named after a river.
   Each data frame have the variables:
\describe{
  \item{name}{site name}
  \item{xkoorlok}{RT90 X coordinate / 10 Used as primary key in SERS}
  \item{xkoorlok}{RT90 Y coordinate / 10 Used as primary key in SERS}
  \item{section}{Name of section (blank if river is not divided in sections)}
  \item{area}{Estimated production area in hektar (per section if the river is divided)}
  \item{used}{logical Indicates if site is used in working group estimations}
  \item{since}{numerical First year site should be used in estimations. Zero = used since start}

}

An object of class \code{character} of length 15.

An object of class \code{character} of length 46.
}
\source{
\url{https://www.slu.se/institutioner/akvatiska-resurser/databaser/elfiskeregistret/}

\url{https://www.ices.dk/community/groups/Pages/WGBAST.aspx}

\url{https://www.ices.dk/community/groups/Pages/WGNAS.aspx}
}
\usage{
efish_sites

WGBAST_rivers

WGNAS_rivers
}
\description{
Dataset DCF::efish_sites contains information about all electrofishing
sites monitored in the DCF-program. Xkoorlok and ykoorlok is the primary key
to find a unique site in SERS. `DCF::WGBAST_rivers` and
`DCF::WGNAS_rivers` are character vectors with rivers monitored by WG.
These vectors can, for example, be used to loop do some processing for all
rivers for a working group.
}
\keyword{datasets}
