% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_pkeys}
\alias{get_pkeys}
\title{Construct a string suitable to use in a where clause to select DCF sites for one river.}
\usage{
get_pkeys(
  river,
  table_prefix = "a.",
  sites = DCFsers::dcf_known_efish_sites(river)
)
}
\arguments{
\item{river}{name of the river to construct a where clause for}

\item{sites}{a data frame with xkoorlok and ykoorlok}

\item{tabe_prefix}{a string used as a table prefix (default = ".a")}
}
\value{
a character string with primary keys for a specific river
}
\description{
Construct a string suitable to use in a where clause to select DCF sites for one river.
}
\examples{
pkeys <- get_pkeys("Rickleån")
}
