% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/allFunctions.R
\name{getUnmatched}
\alias{getUnmatched}
\title{Return a list of countries that are not plotted in corresponding map}
\usage{
getUnmatched(data, varname.x)
}
\arguments{
\item{data}{A dataset with countries as cases}

\item{varname.x}{The variable name for the column that holds the countries names}
}
\description{
Return a list of countries that are not plotted in corresponding map
}

