% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riverinfo}
\alias{riverinfo}
\title{River information}
\format{
riverinfo: A data frame with information about rivers monitored in the DCF-program. Used to translate river names to catchment names.
\describe{
 \item{river}{River name}
 \item{haroname}{Catchment name}
 \item{haronr}{Catchment number}
 }
}
\usage{
riverinfo
}
\description{
River information
}
\keyword{datasets}
