% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{sers_connect}
\alias{sers_connect}
\title{Connect to the SERS database}
\usage{
sers_connect(
  dbname = "fishdb",
  host = "193.10.96.167",
  port = 5432,
  user = "readonly",
  password = "readonly",
  use_latin1 = FALSE
)
}
\arguments{
\item{dbname}{Name of the database containing SERS (default = fishdb)}

\item{host}{Name or ip-number of host with database}

\item{port}{database port number}

\item{user}{database user}

\item{password}{database password}

\item{use_latin1}{set to TRUE if the client uses Latin1 (default = FALSE)}
}
\value{
a connection object
}
\description{
Connect to the SERS database
}
\examples{
\dontrun{
sers <- sers_connect()
res <- dcf_get_efish_data(sers, "Vindelälven", species = "Lax", year = 2018)
}

}
