% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{sers_latest_regdatum}
\alias{sers_latest_regdatum}
\title{Latest SERS update date}
\usage{
sers_latest_regdatum(con)
}
\arguments{
\item{con}{a valid connection to SERS}
}
\value{
The latest registration date in table sers.elfisken
}
\description{
Latest SERS update date
}
\examples{
\dontrun{
con <- sers_connect()
latest_date <- sers_latest_regdatum(con)
}
}
