% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{sers_sites_sf}
\alias{sers_sites_sf}
\title{Return the spatial point of all (DCF) electrofishing point for a river}
\usage{
sers_sites_sf(con, river, sites = DCFsers::dcf_known_efish_sites(river))
}
\arguments{
\item{con}{a database connection to sers (see sers_connect())}

\item{river}{character string with river name}

\item{sites}{dataframe with site info. Must have cols xkoorlok and ykoorlok in SERS-format.}
}
\value{
a SpatialPointsDataFrame with elecrofishing sites

a SpatialPointsDataFrame with elecrofishing sites
}
\description{
Return the spatial point of all (DCF) electrofishing point for a river
}
\examples{
foo <- sers_sites_sp("Emån")
foo

Return the simple feature collection (sf points) of all (DCF) electrofishing point for a river

\dontrun{
sers <- sers_connect()
foo <- sers_sites_sf(sers, "Emån")
foo
}

}
