% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toISO3}
\alias{toISO3}
\title{Turn country names into ISO 3166-1 alpha-3 codes.}
\usage{
toISO3(names)
}
\arguments{
\item{names}{a vector of country names}
}
\value{
a corresponding vector of ISO 3166 codes.  \code{NA} when no match
can be found for an element of \code{names}.
}
\description{
ISO 3166-1 alpha 3 codes are three-letter codes for countries, as specified
by the International Standards Organization.
}
\details{
This function will typically be used with \code{mutate} or \code{transform}
to create a column that is standard across datasets.

TO DO: Add in options for ISO2, num, etc.
}
\examples{
toISO3( c("Russia","China", "Luxembourg", "Nowhere Land"))
}

