% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_add_bundle.R
\name{dcf_add_bundle}
\alias{dcf_add_bundle}
\title{Adds a Bundle Project}
\usage{
dcf_add_bundle(
  name,
  project_dir = ".",
  source_files = NULL,
  open_after = interactive()
)
}
\arguments{
\item{name}{Name of the bundle}

\item{project_dir}{Path to the Data Collection Framework project.}

\item{source_files}{Character vector of paths to standard files form source projects.}

\item{open_after}{Logical; if \code{FALSE}, will not open the project.}
}
\value{
Nothing; creates default files and directories.
}
\description{
Establishes a new data bundle project, used to prepare outputs from standardized datasets.
}
\section{Project}{


Within a bundle project, there are two files to edits:
\itemize{
  \item \strong{\code{ingest.R}}: This is the primary script, which is automatically rerun.
    It should store raw data and resources in \code{raw/} where possible,
    then use what's in \code{raw/} to produce standard-format files in \code{standard/}.
    This file is sourced from its location during processing, so any system paths
    must be relative to itself.
  \item \strong{\code{measure_info.json}}: This is where you can record information
    about the variables included in the standardized data files.
    See \code{\link{dcf_measure_info}}.
}
}

\examples{
project_dir <- paste0(tempdir(), "/temp_project")
dcf_init("temp_project", dirname(project_dir))
dcf_add_bundle("bundle_name", project_dir)
list.files(paste0(project_dir, "/data/bundle_name"))

}
