% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_add_bundle.R
\name{dcf_add_bundle}
\alias{dcf_add_bundle}
\title{Adds a Bundle Project}
\usage{
dcf_add_bundle(
  name,
  project_dir = ".",
  source_files = NULL,
  open_after = interactive()
)
}
\arguments{
\item{name}{Name of the bundle}

\item{project_dir}{Path to the Data Collection Framework project.}

\item{source_files}{A list or character vector, with names as paths to standard files
form source projects (relative to the project's data directory), and distribution file
names as entries. This associates input with output files, allowing for calculation of
a source state, and metadata inheritance from source files.}

\item{open_after}{Logical; if \code{FALSE}, will not open the project.}
}
\value{
Nothing; creates default files and directories.
}
\description{
Establishes a new data bundle project, used to prepare outputs from standardized datasets.
}
\section{Project}{


Within a bundle project, there are two files to edits:
\itemize{
  \item \strong{\code{build.R}}: This is the primary script, which is automatically rerun.
    It should read data from the \code{standard} directory of source projects,
    and write to it's own \code{dist} directory.
  \item \strong{\code{measure_info.json}}: This should list all non-ID variable names
    in the data files within \code{dist}. These will inherit the standard measure info
    if found in the source projects referred to in \code{source_files}.
    If the \code{dist} name is different, but should still inherit standard measure info,
    a \code{source_id} entry with the original measure ID will be used to identify the original
    measure info.
    See \code{\link{dcf_measure_info}}.
}
}

\examples{
project_dir <- paste0(tempdir(), "/temp_project")
dcf_init("temp_project", dirname(project_dir))
dcf_add_bundle("bundle_name", project_dir)
list.files(paste0(project_dir, "/data/bundle_name"))

}
