% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_add_source.R
\name{dcf_add_source}
\alias{dcf_add_source}
\title{Adds a Source Project}
\usage{
dcf_add_source(
  name,
  project_dir = ".",
  open_after = interactive(),
  use_git = TRUE,
  use_workflow = FALSE
)
}
\arguments{
\item{name}{Name of the source.}

\item{project_dir}{Path to the Data Collection Framework project.}

\item{open_after}{Logical; if \code{FALSE}, will not open the project.}

\item{use_git}{Logical; if \code{TRUE}, will initialize a git repository.}

\item{use_workflow}{Logical; if \code{TRUE}, will add a GitHub Actions workflow.}
}
\value{
Nothing; creates default files and directories.
}
\description{
Establishes a new data source project, used to collect and prepare data from a new source.
}
\section{Project}{


Within a source project, there are two files to edits:
\itemize{
  \item \strong{\code{ingest.R}}: This is the primary script, which is automatically rerun.
    It should store raw data and resources in \code{raw/} where possible,
    then use what's in \code{raw/} to produce standard-format files in \code{standard/}.
    This file is sourced from its location during processing, so any system paths
    must be relative to itself.
  \item \strong{\code{measure_info.json}}: This is where you can record information
    about the variables included in the standardized data files.
    See \code{\link{dcf_measure_info}}.
}
}

\examples{
project_dir <- paste0(tempdir(), "/temp_project")
dcf_init("temp_project", dirname(project_dir))
dcf_add_source("source_name", project_dir)
list.files(paste0(project_dir, "/data/source_name"))

}
