% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_build.R
\name{dcf_build}
\alias{dcf_build}
\title{Run a Project's Build Process}
\usage{
dcf_build(project_dir = ".")
}
\arguments{
\item{project_dir}{Path to the Data Collection Framework project to be built.}
}
\value{
A version of the project report, which is also written to
\code{project_dir/docs/report.json.gz}.
}
\description{
Build a Data Collection Framework project,
which involves processing and checking all data projects.
}
\examples{
project_file <- "../../../pophive"
if (file.exists(project_file)) {
  report <- dcf_build(project_file)
}
}
