% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_build.R
\name{dcf_build}
\alias{dcf_build}
\title{Run a Project's Build Process}
\usage{
dcf_build(
  project_dir = ".",
  is_auto = TRUE,
  ...,
  make_diagram = TRUE,
  make_file_log = TRUE
)
}
\arguments{
\item{project_dir}{Path to the Data Collection Framework project to be built.}

\item{is_auto}{Logical; if \code{FALSE}, will run \code{\link{dcf_process}} as if it were run
manually.}

\item{...}{Passes arguments to \code{\link{dcf_process}}.}

\item{make_diagram}{Logical; if \code{FALSE}, will not make a \code{status.md} diagram.}

\item{make_file_log}{Logical; if \code{FALSE}, will not make a \code{file_log.json} output.}
}
\value{
A version of the project report, which is also written to
\code{project_dir/docs/report.json.gz}.
}
\description{
Build a Data Collection Framework project,
which involves processing and checking all data projects.
}
\examples{
project_file <- "../../../pophive"
if (file.exists(project_file)) {
  report <- dcf_build(project_file)
}
}
