% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_data.R
\name{dcf_data}
\alias{dcf_data}
\title{Use Data from a Data Collection Project}
\usage{
dcf_data(
  variables = NULL,
  project = ".",
  data_format = NULL,
  project_type = "bundle",
  ...,
  unify = TRUE,
  only_selected = FALSE,
  cache = tempdir(),
  refresh = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{variables}{A character vector of variable names to be loaded, or a selected
subset of a project data dictionary, as returned from \code{\link{dcf_variables}}.}

\item{project}{Path to a local project, or the GitHub account and repository name
(\code{"{account_name}/{repo_name}"}) of a remote project.}

\item{data_format}{The data format to select, between \code{tall} and \code{wide}.
Useful if there are duplicate measure names between files of different formats.}

\item{project_type}{Project type to select, between \code{bundle} and \code{source}.}

\item{...}{Additional arguments passed to \code{\link{dcf_report}}.}

\item{unify}{Logical; if \code{FALSE}, will return \code{data} as a list with entries for each
file. Otherwise (by default), will attempt to combine the loaded data.}

\item{only_selected}{Logical; if \code{TRUE}, will drop columns that were not included in
\code{variables}, other than ID columns.}

\item{cache}{Path to a directory in which to store downloaded files. Within this directory,
the repository structure will be recreated within an account-named directory.}

\item{refresh}{Logical; if \code{TURE}, will download files even if they exist in the \code{cache}.}

\item{verbose}{Logical; if \code{FALSE}, will not show status messages.}
}
\value{
A list with entries for metadata (the datapackage resource entry for each file loaded)
and data (a tibble or list of tibbles of the unified or separately loaded files).
}
\description{
Load the standard or distribution data from a local or remote data collection project.
}
\examples{
# retrieve the full bundle file that includes the `epic_rsv` measure
bundle <- dcf_data(
  "epic_rsv",
  "dissc-yale/pophive_demo",
  data_format = "tall",
  verbose = FALSE
)
bundle$data

if (require("ggplot2", quietly = TRUE)) {
  # extract short names from data
  labels <- vapply(
    bundle$metadata[[1L]]$schema$fields[[3L]]$info$levels,
    function(measure) measure$info$short_name,
    ""
  )

  # show trends from different measures over time
  bundle$data |>
    dplyr::filter(
      time >= as.Date("2024-01-01"),
      measure != "epic_all_encounters"
    ) |>
    dplyr::mutate(measure = labels[measure]) |>
    ggplot(aes(x = time, y = value_scaled, color = measure)) +
    theme_dark() \%+replace\%
    theme(
      panel.background = element_rect(fill = FALSE, color = FALSE),
      geom = element_geom(fill = "white")
     ) +
    geom_smooth(
      method = "gam",
      formula = y ~ s(x, bs = "cs", k = 50L)
    )
}
}
\seealso{
Other data user interface functions: 
\code{\link{dcf_report}()},
\code{\link{dcf_variables}()}
}
\concept{data user interface functions}
