% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_init.R
\name{dcf_init}
\alias{dcf_init}
\title{Initialize a Data Collection Project}
\usage{
dcf_init(
  name,
  base_dir = ".",
  data_dir = "data",
  github_account = "",
  use_git = TRUE,
  open_after = interactive()
)
}
\arguments{
\item{name}{Name of the source.}

\item{base_dir}{Path to the directory containing sources.}

\item{data_dir}{Name of the directory to store projects in, relative to \code{base_dir}.}

\item{github_account}{Name of the GitHub account that will host the repository.}

\item{use_git}{Logical; if \code{TRUE}, will initialize a git repository.}

\item{open_after}{Logical; if \code{FALSE}, will not open the project.}
}
\value{
Nothing; creates default files and directories.
}
\description{
Establishes a new data collection framework project.
}
\section{Data Collection Project}{


A data collection project starts with a \code{settings.json} file, which
specifies where source and bundle projects live (a \code{data} subdirectory by default).

The bulk of the project will then be in the source and bundle projects, as created
by the \code{\link{dcf_add_source}} and \code{\link{dcf_add_bundle}}.

Once these sub-projects are in place, they can be operated over by the
\code{\link{dcf_build}}, which processes each sub-project using
\code{\link{dcf_process}}, and checks them with \code{\link{dcf_check_sources}},
resulting in a report.
}

\examples{
base_dir <- tempdir()
dcf_init("project_name", base_dir)
list.files(paste0(base_dir, "/project_name"))

}
