% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_variables.R
\name{dcf_variables}
\alias{dcf_variables}
\title{View Project Variables}
\usage{
dcf_variables(project = ".", ...)
}
\arguments{
\item{project}{Path to a local project, or the GitHub account and repository name
(\code{"{account_name}/{repo_name}"}) of a remote project.}

\item{...}{Additional arguments passed to \code{\link{dcf_report}}.}
}
\value{
A tibble containing variables:
\tabular{ll}{
  \code{name} \tab Name of the variable, as it appears in the data file. \cr
  \code{type} \tab The value's storage type. \cr
  \code{n} \tab Number of non-missing observations within the file. \cr
  \code{duplicates} \tab Number of duplicated values within the file. \cr
  \code{missing} \tab Number of missing values within the file. \cr
  \code{file} \tab The file containing the variable; a path relative to the project root. \cr
  \code{short_name} \tab Short name, if included in measure info. \cr
  \code{long_name} \tab Long name, if included in measure info. \cr
  \code{short_decription} \tab Short description, if included in measure info. \cr
  \code{long_description} \tab Long description, if included in measure info. \cr
  \code{unit} \tab Unit (what the value represents), if included in measure info. \cr
  \code{category} \tab The measure's category, if included in measure info. \cr
}
}
\description{
Get a list of measures (a codebook) that are included in a data collection project.
}
\examples{
dcf_variables("dissc-yale/pophive_demo")
}
